//add a few files

+++ linbox-old/src/library/objects/element/element_padic.h
@@ -53,7 +53,7 @@ namespace LinBox{
  
     public:
       std::vector<long int> rep;
-      long int exp;
+      int exp;
     };
 }
 #endif
